+++ app/src/main/java/com/example/myportfolio/data/sampledata/Assets_assets.kt
@@ -1,0 +1,5 @@
+package com.example.myportfolio.data.sampledata
+
+// class Assets_assets {
+// val assets: List<Asset> = listOf()
+// }
