+++ app/src/main/java/com/example/myportfolio/domain/models/Assets.kt
@@ -1,9 +1,3 @@
 package com.example.myportfolio.domain.models
+
 interface Asset
-
-// class Currency(val name: String, val quantity: Int, val price: Double) : Asset
-
-//data class Bond(val name: String, val quantity: Int, val price: Double, val issuer: String, val maturityDate: String) : Asset
-
-//data class Stock(val name: String, val quantity: Int, val price: Double, val symbol: String) : Asset
-
