+++ app/src/main/java/com/example/myportfolio/domain/models/Bonds.kt
@@ -3,8 +3,8 @@ package com.example.myportfolio.domain.models
 data class Bonds(
     val id: Int,
     val name: String,
-    //val quantity: Int,
     val price: Double,
-    val currency:String,
+    val currency: String,
     val issuer: String,
-    val maturityDate: String) : Asset
+    val maturityDate: String
+) : Asset
