+++ app/src/main/java/com/example/myportfolio/domain/models/Stocks.kt
@@ -1,9 +1,11 @@
 package com.example.myportfolio.domain.models
-data class Stocks(val id: Int,
-                  val name: String,
-                 // val quantity: Int,
-                  val price: Double,
-                  val currency: String,
-                  val Industry: String,
-                  val Countryissuer: String,
-                  val symbol: String): Asset
+
+data class Stocks(
+    val id: Int,
+    val name: String,
+    val price: Double,
+    val currency: String,
+    val industry: String,
+    val countryIssuer: String,
+    val symbol: String
+) : Asset
