--- app/src/main/java/com/example/myportfolio/sampledata/Assets_assets.kt
@@ -1,6 +1,0 @@
-package com.example.myportfolio.sampledata
-
-class Assets_assets {
-     val assets = listOf<Assets>()
-
-}
